refactor : enum 패키지 생성
