+++ src/main/java/enums/HttpStatus.java
@@ -1,4 +1,4 @@
-package util;
+package enums;
 
 public enum HttpStatus {
     OK(200, "OK"),
