+++ src/main/java/util/HttpResponse.java
@@ -1,5 +1,8 @@
 package util;
 
+import enums.FileContentType;
+import enums.HttpStatus;
+
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.util.HashMap;
