+++ src/main/java/util/RequestParser.java
@@ -1,5 +1,6 @@
 package util;
 
+import enums.HttpMethod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
